requirements.txt
aniso8601==1.3.0
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Babel==0.11.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
futures==3.2.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
MySQL-python==1.2.5
python-dateutil==2.6.1
python-ldap==2.5.2
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.13
